itunes_store/requirements.txt
MySQL-python==1.2.4c1
SQLAlchemy==0.7.9
transaction==1.3.0
zope.sqlalchemy==0.7.1
# pyyaml, http://pyyaml.org/download/pyyaml/PyYAML-3.10.tar.gz, python setup.py install
